\name{caratREG-package}
\alias{caratREG-package}
\alias{caratREG}
\docType{package}
\title{
\packageTitle{caratREG}
}
\description{
\packageDescription{caratREG}
}
\details{

\packageIndices{caratREG}
}
\author{
\packageAuthor{caratREG}

Maintainer: \packageMaintainer{caratREG}
}
\references{
Ma, W., Tu, F., & Liu, H. (2020). \emph{Regression analysis for covariate-adaptive randomization: A robust and efficient inference perspective}. arXiv preprint arXiv:2009.02287.
}
\keyword{ package }
