% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{tau.interact}
\alias{tau.interact}
\title{Regression with interaction}
\usage{
tau.interact(Y, A, B, X = NULL, pi, conf.level = 0.95)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{B}{a numeric vector of stratum labels. Its length should be the same as
the number of subjects.}

\item{X}{an (optional) numeric design matrix containing additional covariates
used in the regression.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{conf.level}{confidence level of the interval. Default is 0.95.}
}
\value{
A list of class \code{"htest"} containing the following components: 
 \item{statistic}{the value of the t-statistic.}
 \item{p.value}{the p-value for the test.}
 \item{conf.int}{a confidence interval under chosen level \code{conf.level} for the difference
 in treatment effect between treatment group and control group.}
 \item{estimate}{estimated treatment effect difference between treatment
 group and control group.}
 \item{method}{a character string indicating what type of regression was performed.}
}
\description{
Estimating and inferring the treatment effect based on regression with
interaction.
}
\details{
Estimating and inferring the treatment effect based on regression with interaction. It
implements the methods as described in Sections 3.3 and 4.3, Ma et al. (2020).
}
\examples{
#The code replicates the simulation setting of Model 2 in Section 6, Ma et al. (2020).
n <- 1000
pi <- 0.5
X1 <- rgamma(n,2)
X2 <- sample(c(1,2,3),n,replace = TRUE, prob = c(0.3,0.6,0.1))
X3 <- rpois(n,3)
X4 <- rbeta(n,2,2)
X1_S <- ifelse(X1 < 2.5, 1, 2)
B <- as.numeric(interaction(X1_S,X2))
X <- cbind(X1,X3)
A <- sample(c(0,1),n,replace=TRUE,prob=c(1-pi,pi))
alphavec <- c(5,8,3,12)
Y0 <- alphavec[1]*X1+log(alphavec[3]*X1*log(X3+1)+1)+alphavec[4]*exp(X4)+rnorm(n,sd = 2)
Y1 <- 1.483708+alphavec[2]*X2^2+log(alphavec[3]*X1*log(X3+1)+1)+rnorm(n,sd = 1)
Y <- Y0*(1-A)+Y1*A
tau.interact(Y, A, B, X, pi)
}
\references{
Ma, W., Tu, F., & Liu, H. (2020). \emph{Regression analysis for
 covariate-adaptive randomization: A robust and efficient inference
 perspective}. arXiv preprint arXiv:2009.02287.
}
